fix: hocs components
