+++ src/components/Calendar/types.ts
@@ -1,12 +1,3 @@
-export interface CalendarProps {
-  showCalendar: boolean;
-  showClearButton: boolean;
-  activeDay: number | null;
-  rangeEndDay?: number | null;
-  onClear: () => void;
-  onClickCell: (timestamp: number) => void;
-}
-
 export interface CalendarContainerProps {
   $showClearButton: boolean;
   $showCalendar: boolean;
