+++ src/types/calendar/CalendarProps.ts
@@ -1,0 +1,8 @@
+export interface CalendarProps {
+  showCalendar: boolean;
+  showClearButton: boolean;
+  activeDay: number | null;
+  rangeEndDay?: number | null;
+  onClear: () => void;
+  onClickCell: (timestamp: number) => void;
+}
