+++ src/types/calendar/index.ts
@@ -1,0 +1,1 @@
+export { type CalendarProps } from './CalendarProps';
