+++ src/types/index.ts
@@ -1,3 +1,4 @@
+export * from './calendar';
 export * from './config';
 export * from './day';
 export * from './hooks';
